Added a hint about what to do with special words
This provides no insight at all:
Certain words are only capitalized if they are the first word in the title: "a", "and", "the", and "of".

Many students have no idea they need to create a variable to reference.  So I'm sure telling them to use `include?` on some array would likely confuse many to think since we're splitting the title, that that's what you should call it on.