Real World Example for .. in Loops
I am not sold on how I am presenting the relevance of not knowing how long the line is. 
I am also not sold on using child as the thing we're iterating through because it doesn't quite map to properties of an object, but for each as a concept is not limited to objects. 

Happy to have feedback or change the example per your ideas! :)